Added basic mailto, see notes
If you are having trouble getting the email link to work (I did), go into your browser setting and confirm that email handlers are not blocked.

Then go to your email client and enable it as a default handler.

I had no handler enabled - this took a while to troubleshoot.